+++ autogpt_platform/frontend/src/lib/autogpt-server-api/client.ts
@@ -909,7 +909,6 @@ export default class BackendAPI {
       graph_version: graphVersion,
     });
   }
-
 }
 
 /* *** UTILITY TYPES *** */
